fusion-dr-predicator/requirement.txt
streamlit
torch
torchvision
pillow
numpy
matplotlib
seaborn
plotly
